+++ packages/edit-site/src/components/page-templates/index-legacy.js
@@ -133,6 +133,7 @@ export default function PageTemplates() {
 		<Page
 			className="edit-site-page-templates"
 			title={ __( 'Templates' ) }
+			ariaLabel={ __( 'Templates' ) }
 			actions={ <AddNewTemplate /> }
 		>
 			<DataViews