+++ packages/edit-site/src/components/post-list/index.js
@@ -273,6 +273,7 @@ export default function PostList( { postType } ) {
 	return (
 		<Page
 			title={ labels?.name }
+			ariaLabel={ labels?.name ?? '' }
 			actions={
 				<>
 					{ labels?.add_new_item && canCreateRecord && (