+++ routes/pattern-list/stage.tsx
@@ -258,6 +258,7 @@ function PatternList() {
 	return (
 		<Page
 			title={ __( 'Patterns' ) }
+			ariaLabel={ __( 'Patterns' ) }
 			subTitle={ __(
 				'Reusable design elements for your site. Create once, use everywhere.'
 			) }