+++ routes/styles/stage.tsx
@@ -74,6 +74,7 @@ function Stage() {
 			}
 			className="routes-styles__page"
 			title={ __( 'Styles' ) }
+			ariaLabel={ __( 'Styles' ) }
 		>
 			<GlobalStylesUIWrapper
 				path={ section }