+++ routes/template-list/stage-legacy.tsx
@@ -272,6 +272,7 @@ function TemplateListLegacy() {
 	return (
 		<Page
 			title={ __( 'Templates' ) }
+			ariaLabel={ __( 'Templates' ) }
 			className="template-page"
 			actions={ <AddNewTemplate /> }
 			hasPadding={ false }